deadlock to correct

+++ philo/utils.c
@@ -6,13 +6,13 @@
 /*   By: cogata <cogata@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/31 14:30:47 by cogata            #+#    #+#             */
-/*   Updated: 2024/06/05 15:00:32 by cogata           ###   ########.fr       */
+/*   Updated: 2024/06/08 11:07:15 by cogata           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philosophers.h"
 
-void error_exit(char *error)
+void	error_exit(char *error)
 {
 	printf("%s\n", error);
 	exit(1);
@@ -26,11 +26,11 @@ int	ft_strcmp(char *s1, char *s2)
 	while (s1[i] != '\0')
 	{
 		if (s1[i] != s2[i])
-			return ((unsigned char) s1[i] - (unsigned char) s2[i]);
+			return ((unsigned char)s1[i] - (unsigned char)s2[i]);
 		i++;
 	}
 	if (s1[i] != s2[i])
-		return ((unsigned char) s1[i] - (unsigned char) s2[i]);
+		return ((unsigned char)s1[i] - (unsigned char)s2[i]);
 	else
 		return (0);
 }
@@ -47,12 +47,12 @@ size_t	ft_strlen(const char *s)
 	return (i);
 }
 
-long	ft_long_atoi(const char *nptr)
+size_t	ft_long_atoi(const char *nptr)
 {
-	char		*nbr;
-	int			i;
-	int			sign;
-	long int	res;
+	char	*nbr;
+	int		i;
+	size_t	sign;
+	size_t	res;
 
 	nbr = (char *)nptr;
 	i = 0;
